Split FormLoaderContainer into three components for better encapsulation
